[RTCInsertableStreams] Fix timestamp field for RTCEncodedVideoFrame
Use the RTP Timestamp instead of the received time.
This has the advantage that it has the same value in both sender and receiver.
Previously, we were using ReceivedTime, which only worked on the receiver
and was always zero in the sender.

Bug: 1052765
Change-Id: Ib6bd251f2d14aefeabd0d0f315955794865e132f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2098627
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749821}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 99d1f3937bb684a05e00d562371677d064edf276