Slightly delay rendering for font-loading via JS API
As discussed in the design doc (https://bit.ly/36E8UKB), if we start
loading a font via the Font Loading API, we should also slightly delay
rendering for it so that it has a higher change to meet the first
paint. In other words, document.fonts.load is treated the same as
<link rel="preload">.

This patch implements that.

Bug: 1040632
Change-Id: I1d21b8130c705feb87b5f1147eba8743820052e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095412
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749841}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dff8b1539bfdecf1474d126aff99ab8ce0c42f23