IndexedDB: Only prioritize idb tasks if page is visible
Previously for the HighPriorityDatabaseTaskType experiment, all idb
tasks were set to high priority. This, however, can cause performance to
degrade on the foreground tab if sharing a renderer with background
pages that are runnnig a lot of idb tasks. This CL changes it so that
idb tasks are only prioritized for visible pages.

Bug: 984374
Change-Id: If8296517f27111decb265a1d79aa2d49e734a9d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100107
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749905}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de1e694e8101689bd6a8a163ed0c5545b0b61182