Replace the native menu control with the internal popup based one on Mac.
The internal popup is going to be used for select when
UseCommonSelectPopup feature is enabled.

The internal select popup is going to be consistent across Linux,
Windows and Mac in order to provide the same appearance,
styling capabilities and behavior.

Bug: 1009905
Change-Id: Iac07b58df21e6e6ccdee43da3df44f3646696639
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088228
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#750022}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f409e632b923eadd7342485b3ab0829bdd3584f2