base/android: Don't instrument the linker code for orderfile builds.
The orderfile build instruments all functions. However, some code is in
a different .so, namely the linker code. As a consequence, linking the
library fails as the instrumentation functions are not found.

To fix this, make sure that the code is not instrumented.

Bug: 1061359
Change-Id: I7b308e6beb73a42d47f70f8116ef28078b18d0df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100942
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: David Turner <digit@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750096}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f044063a3156d76c1cbdc8d4ee5de0761e5bee15