Adds missing awaits to DevTools layout tests
There have been a number of flakes in these tests (and more besides). In
this particular case, though, the tests are suspected to fail because of
a race condition where the showView Promise resolves sometimes, and not
in others. This CL adds the appropriate awaits.

Bug: 1046784
Change-Id: Ic24b0a1aad629d2a45899922af192e79a6c1dd24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095010
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750151}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd83752630361ab71bb158f69cf5b376cef23e38