[Offline indicator v2] Fix blink layout height and fullscreen
Today, most parts of the code assume the browser controls min-height
is 0. This causes issues with the layout size and the fullscreen mode.
The most obvious issues currently are:
- If a min-height is set, ChromeFullscreenManager#controlsResizeView()
always returns true since the controls aren't completely hidden.
- Similarly if a min-height is set, position: fixed elements are sized
as if the controls are fully showing.

This CL changes the "browser controls resize/shrink blink" logic to try
and fix the issues listed above. With the changes in this CL, the
controlsResizeView bit will be false if the controls are at the
min-height.

Bug: 1049301
Change-Id: Ieb95e92c983b56b42a640f726c60c2eec9964064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2048276
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750260}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 122f7ac1a78a67a356299b9f2bfcefa4f22e36c4