Refactoring sharedGpuContextTest to stop using AcceleratedUsage
Replacing instances of AcceleratedResourceUsage for SharedImage and
Bitmap resource providers.
Changing expectation of a test as this now should fail and not fallback
into the Bitmap mode unless we explicitly call it.
Adding a new SharedGPUContextTest with the Viz context system, to be
able to validate the correct creation of a CanvasResourceProvider
of SharedImage.

Also modifying gpu_test_utils to set the context to not lost.

Bug: 1035589, 1021287
Change-Id: I15342b57f3721bad7bb2b44af184c5b0ae8a4764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095878
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750301}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c62c1140a36d7c5231f3f6b36d0089d6080eecd6