Stop always pixel-snapping clip rects placed in ClipPaintPropertyNode
Also change GeometryMapper to use unsnapped clip rects.

Such pixel-snapping introduces rounding errors when using
GeometryMapper for non-paint-output use-cases such as
IntersectionObserver.

Other previous call sites to ClipRect() that have to do with paint
or compositing are converted to PixelSnappedClipRect().

This CL stores two clip rects on each ClipPaintPropertyNode: the
unsnapped one and the snapped one (if applicable; otherwise the
snapped one is the same as the unsnapped one).

We need to store both clip rects because not all clips should be
pixel-snapped even in paint. An example of clips that should not
be pixel-snapped is in SVG, which does not have pixel snapping.

Bug: 992765

Change-Id: I92dc5f3f698f1d9ae2fd64a0c08cda323a8d8f6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092698
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750329}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6f04efdc6f3fa1483926882528b3104707fa20a