Eliminate redundant calls of Font::Update() in FontFaceSetDocument
Both call sites of Font::Update() in FontFaceSetDocument are
redundant, so this patch removes them.

The first call site is redundant as StyleResolver::ComputeFont()
will set the underlying Document's FontSelector to the font.

The second call site is redundant as it repeatedly sets the Document's
FontSelector to the font.

This is also preparation for eliminating Font::Update().

Bug: 1049295
Change-Id: I43196beb8c77e420083a125a6050d73d06e7b9c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103146
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750591}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 49a124c296001726d66963f36c838c645f8db322