Implement mid-ligature selection for LayoutNG
This patch adds mid-ligature support when painting selection.

This was originally implemented in issue 473476, but it only
supported legacy painters. We overlooked this feature in
LayoutNG as there were no tests, other than minor pixel
differences.

This patch, in addition to implement it, adds 3 pixel tests:
  virtual/text-antialias/selection/select-ligature-1.html
  virtual/text-antialias/selection/select-ligature-2.html
  virtual/text-antialias/selection/select-ligature-3.html
for when selecting the first, last, and the middle character
of a ligature.

One test that fails in legacy:
  external/wpt/css/css-pseudo/selection-text-shadow-016.html
starts to fail in LayoutNG too, because without issue 473476,
selected text accidentally overdraw the `text-shadow` that
should not be rendered. This is tracked in issue 932343.

Bug: 473476, 932343, 1025341
Change-Id: Iee2b795e96c1f6a8d07683c57c13a4cdbec88832
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103968
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750701}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3fd48e73607c42cf77ed52bac33664c3f2b468e3