Worker: Remove OffMainThreadWorkerScriptFetchOption
This option has been used for controlling script fetch behavior for
dedicated, shared, and service workers. Now off-the-main-thread worker
script fetch is enabled by default for shared workers and service
workers, and this option takes effect only on dedicated workers.

This option is replaceable with a combination of the flag of
PlzDedicatedWorkers and the script type of dedicated workers, and that's
more preferable in terms of code simplicity. This CL replaces the option
with the combination, and removes the option.

Bug: 835717
Change-Id: I81c8f7be4663353471c56624fc16569e2aada156
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2104568
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750712}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e13bda2dfd1f8a97aebd90d5a74593d757b10a4