Restore the event delegate for a CSSTransition after a call to animation.effect
Transition events were dropped after a call to animation.effect.
CSSTransitions uses an event delegate to dispatch events as a result
of changes to the timing phase.  The event delegate was getting
dropped when the effect changed. A special case is when effect is set
to null. In this case, we need to dispatch a transitioncancel prior
to losing the old event delegate.

Bug: 1059936
Change-Id: I9eda3484fb0c9de08e8398e1975031d6a1f58940
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103034
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750775}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e6bca36e2f068c623e3bd9e24b1f38b20062665e