Referrer-WPT: Fix weird referrer test timing issue
This CL fixes a very tricky timing issue I grappled with debugging:
I noticed that a non-deterministic amount of the worklet-referrer tests
always failed in Chromium when run from the wpt.live site, e.g.,
https://wpt.live/worklets/paint-worklet-referrer.https.html. This was
NOT reproducible locally via any of the following:
 a.) `./wpt serve` and running the tests in Chromium
 b.) run_web_tests.py to run them automatically with Chromium

I made a dummy WPT PR [1] just to trigger a wpt.live deployment so I
could debug my changes in the wpt.live environment, where the tests were
failing for some reason. The main referrer test window opens up a bunch
of child windows which run a test, and report their results back to the
main window IN ORDER. The test were failing for two reasons:
 - assert_equals: expected "LOADED" but got "RESOLVED"
 - assert_equals: expected (string) "RESOLVED" but got (object) object "[object Object]"

This indicated some type of timing issue between the child windows, and I
suspected that one window was reporting back its results when the main
window was expecting a different child's results. I dug into the second
assertion failure above, logging the "object". It was a timeout object
automatically reported by the testharness.

This made me think that the child windows were being kept open too long
after their test finished, and automatically reported an "extra" timeout
object back to the main window, when it was expecting a later child
window's results. I finally realized that on line 3, there was a call
`add_completion_callback(() => win.close())`, the intention being to
close a child window after its test completed.

I found in the testharness API docs, that add_completion_callback doesn't
call the callback after the test completes, but only after *all* tests
complete [2]. This means all child windows were being kept open for the
entire duration of the main test. This caused some child windows to
report the timeout object to the main window when the server is slow, like
on the wpt.live site, but never was reproduced.

The fix is to not use add_completion_callback. We could instead use
add_result_callback which would work, but it is global, so it would end
up attempting to invoke window.close O(n^2) times for O(n) tests (I've
confirmed this). The best solution is to plumb each promise_test object
through, and add a cleanup callback that closes the child window. This
CL implements that. With this fixed, it might save a lot of debugging
hours in the future \o/.

[1]: https://github.com/web-platform-tests/wpt/pull/22269
[2]: https://web-platform-tests.org/writing-tests/testharness-api.html#callback-api

R=nhiroki@chromium.org

Bug: N/A
Change-Id: Icac7e58a5bfc218a86549f11804b2bc856b815f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103999
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750878}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97206802dfa566ce6a8c30e04b00ffb913668c40