Don't call Font::Update() with explicit nullptr
We want to eliminate Font::Update(), as it allows switching the
FontSelector of a Font, which is a bad pattern.

This patch eliminates all call sites that calls the function with an
explicit nullptr, by replacing it with on-demand creation of
Font::font_fallback_list_ with a null FontSelector.

Note that:
- We can't DCHECK non-null FontSelector for Font::Update(). Some callers
  may call it with a nullable parameter
- We can't create a FontFallbackList on all Font objects. Otherwise, we
  may get a memory regression. So we use on-demand creation.

Bug: 1049295
Change-Id: Icc1e87c8517220a2ff8d92f00b311546e8e77541
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106394
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751013}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 40fc23688ba6d7eb84d6ed2d525d4c14596c8e74