[FragmentItem] Fix text snap-to-pixel behavior for culled inline box
This patch modifies the snap-to-pixel behavior implemented in
r736251 <crrev.com/c/2024133>.

The CL changed to snap the top of the padding box of the
parent box to match legacy and |NGPaintFragment| behavior.
It did so even for culled inline boxes too, but this part
was not compatible.

This patch changes the logic to ignore culled inline boxes
for the snap-to-pixel behavior, matching legacy and
|NGPaintFragment|.

Bug: 982194
Change-Id: I667f35d4b55909505502633b2d7e70b058191c8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106044
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751062}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec2ca7ddadeb4398077060d2232c231cc3aaeb36