Check IsDetached in IsARIAControlledByTextboxWithActiveDescendant
Added AXObject::IsDetached to check that the ax_object_cache_ is still
valid in AXObject::IsARIAControlledByTextboxWithActiveDescendant to
fix a crash due to nullptr dereferencing on the ax_object_cache_.

The crash happens in the following scenario, when there is a listbox
options dropdown where Javascript dynamically remove/add the dropdown
option nodes. When Windows Narrator's focus is set on the text content
of the option, and we call Javascript function hide(), a crash could
occur in some cases due to the following (I was unable to repro this
with NVDA):
 1. Javascript remove() nodes will trigger AXObject::Detach(), which
    sets ax_object_cache_ to nullptr.
 2. Meanwhile, Windows Narrator's focus remains on the listbox option,
    and triggers AXNodeObject::OnNativeFocusAction(), which calls
    AXObject::IsARIAControlledByTextboxWithActiveDescendant() and
    retrieves nullptr for ax_object_cache_ and we call FocusedObject()
    on the nullptr cache and causes an exception.

<button onclick="hide()">hide dropdown</button>
<div role="listbox">
  <div role="option" tabindex="1" id="0">option 1</div>
  <div role="option" tabindex="1" id="1">option 2</div>
  <div role="option" tabindex="1" id="2">option 3</div>
</div>

<script>
  function hide() {
    document.querySelector('div[role="listbox"]').remove();
  }
</script>

For scenario similar to the example above, I had difficulty to have
browsertests/unittests to exercise this scenario.

Bug: 1021740
Change-Id: Ib86db0df64c230d341add63faf28a582792fa8c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107429
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751106}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d9fefc87b9eed82040e1eb023513aaf8d75cb63