Use gpuBenchmarking for wheel scroll in wheel event tests
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. We should replace
eventSender.continuousMouseScrollBy with gpuBenchmarking.smoothScrollBy
in the wheel event tests. In the wheel event tests, there is a check for
wheelDeltaX or wheelDeltaY, which has a multiplier factor of 3, which I
think should change to kTickMultiplier/MouseWheelEvent::kWheelDelta,
120/53.

Bug: 1047176
Change-Id: I534cbd25f2a9c84f6f5d891742438facdcda91ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2084091
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751210}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 92439e8f84da949e91656aebaf86548659a3d6a1