heap: Prohibit Vector::Fill for on-heap vectors.
Vector::Fill uses std::fill, which I believe is not safe for concurrent
marking. Currently there are no usages of Vector::Fill for on-heap
vectors.

Disabling Vector::Fill for on-heap vectors will give a compile time
error if anyone tries to use it. I think we can avoid introducing an
atomic version of Vector::Fill and revisit it if and when such a usage
is wanted.

Bug: 986235
Change-Id: I7ac1cb95be3456d86e75f39e0aaf27120e388a84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106057
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751340}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ca7266dadca4e4e10c6a43d9503df8b4bbb29e22