Expose blink accessibility objects hidden via display:none on ancestor
Accessible objects styled with display:none are supposed to be exposed
when they have an id. However, this currently only works if the
display:none style targets the element directly. It also needs to work
if the display:none targets an ancestor.

This is also a necessary precursor to
https://chromium-review.googlesource.com/c/chromium/src/+/2090736

Bug: 651614
Change-Id: I789cda53f303ddfc3304f6109102b71090619aa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095880
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Adam Ettenberger <Adam.Ettenberger@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#751439}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 283a1a3bebeeb2af40949aa211f2339d34db371f