Eliminate a redundant parameter from FontBuilder::CreateFont{ForDocument}
For all callers of FontBuilder::CreateFont() and CreateFontForDocument()
the value of |font_selector| is always equal to the FontSelector of the
current document.

Hence, this patch removes it. It also eases the elimination of the
deprecated function Font::Update().

Bug: 1049295
Change-Id: I1e5521c00191a9af0f670cc985c1de198734bece
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107696
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751488}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dfb591550c6e30ed1d2ce6ea6950413cbcf2c273