Re-enable and re-baseline tests affected by asynchronous source mapping
Two of the tests need a new baseline due to timing changes. The rest
can simply be re-enabled. We will keep an eye on the flakiness
dashboard to ensure that the timing is stable for the re-baselined
tests.

R=petermarshall@chromium.org

Bug: chromium:1032016
Change-Id: I9dc199d74e1cedcaa42722483c5d3abb8c75bbdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109770
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751642}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b76d789f5af85f5a52863ebf36a39e4b8f3e43b0