SVG: Fix clipPath for positioned <foreignObject>
In this scenario, the x & y attributes are baked into both the the
content-rect of the <foreignObject> (as passed to the clipper), *and*
the transform.  As a result, any x & y offsets were being double-applied
to a <clipPath> that used clipPathUnits="objectBoundingBox".

We modify |ReferenceBoxForEffects| so that it removes the unnecessary
transformation from the reference box when it's called on an SVG foreign
object.

Bug: 1009222
Change-Id: Ifc009eb84d46d49c7840e2903d44a90502d6b51d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101320
Commit-Queue: Ian Prest <iapres@microsoft.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751792}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b585a9424bb1c7aa5d7c913d8650f34d2ebf0a0