Add support for animationcancel
The animationcancel event was missing from the set of supported events
for a CSSAnimation.  This patch adds support for the missing event.  The
event includes an elapsed time member for the animation, which is also
plumbed through in this patch.

With support for animationcancel, most of the remaining test timeouts
have been addressed for CSSAnimations. The next step is to handle
reattachment of an event delegate after a call to animation.effect.

Spec:
https://drafts.csswg.org/css-animations-1/#eventdef-animationevent-animationcancel


Bug: 1059968
Change-Id: I83714a06c328a77dc4969928c2c926f3cb207fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107302
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752087}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 49981cfdf89f56e9c82f7848df7daa702d46085f