SubtreeVisibility: Ensure scrollIntoView scrolls even when hidden.
The spec for scrollIntoView does not say anything specific about
principal boxes not existing or scrollIntoView not acting on hidden
content. So, this restores the behavior similar to `visibility: hidden`:
we scroll to the element even though it's hidden.

Note that it's safe to do this, since layout is already forced for this.

Also note that anchor link navigation follows the same path, so the test
expectation for that is updated as well.

R=chrishtr@chromium.org

Change-Id: I6079a4fa396ddb768c17d74e47fd7b72bd23f99f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112142
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752111}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 822ccdfc57e3afa42b74e6ec653f0b5acc341f02