Rename PaintPhase::kSelection to kSelectionDragImage
The old name was confusing with the common meaning of selection
for which we paint selection tints in the foreground paint phase.
The new name makes it clear the paint phase is for the selection's
drag image, not the common meaning of selection.

Bug: 1061070
Change-Id: I762623bea3f679cc7582c542833071475eb54b55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111255
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752124}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d32334fbabd1285d56dc24b27002f503c82395ca